refactor: split utils/validation.py into domain-specific modules (#21)
Split the 247-line utils/validation.py file into focused domain modules:

**New structure:**
- utils/validation/memory.py - Memory key, data, TTL, tags validation
- utils/validation/documents.py - Document content, metadata, search validation
- utils/validation/common.py - Email, URL, sanitization, JSON schema validation
- utils/validation/__init__.py - Package exports for backward compatibility

**Benefits:**
- ✅ Domain separation - Related validation functions grouped together
- ✅ Discoverability - Easy to find validation for specific domains
- ✅ Maintainability - Smaller, focused files easier to understand
- ✅ Backward compatibility - All functions still available at package level
- ✅ Testing - Domain-specific validation can be tested in isolation

**Testing:**
- Added comprehensive test suite in test_validation_package.py
- All 300 tests pass (290 existing + 10 new validation package tests)
- Verified modular imports work correctly
- Confirmed backward compatibility maintained

Addresses GitHub Issue #21.
